new files

+++ include/CLargestRect.h
@@ -1,6 +1,8 @@
 #ifndef CLargestRect_H
 #define CLargestRect_H
 
+#include <iostream>
+
 // Template class to extract largest rectange from array of values DATA of type VALUE
 //
 // DATA must support the getValue(int ix, int iy) method.
@@ -55,15 +57,15 @@ class CLargestRect {
 
   Rect largestRect(const VALUE &value) const {
     Point best_ll( 0,  0);
-    Point best_ur(-1, -1); // bug fix, was 0,0
+    Point best_ur(-1, -1); // bug fix, was 0, 0
 
     std::vector<int> c;
 
-    c.resize(size_t(width_ + 1));
+    c.resize(uint(width_ + 1));
 
     int width1, x0, w0;
 
-    using Xw = std::pair<int,int>;
+    using Xw = std::pair<int, int>;
 
     Xw xw;
 
@@ -75,7 +77,7 @@ class CLargestRect {
       width1 = 0;
 
       for (int x = 0; x < width_ + 1; ++x) {
-        auto xi = size_t(x);
+        auto xi = uint(x);
 
         if      (c[xi] > width1) {
           s.push_back(Xw(x, width1));
@@ -89,7 +91,7 @@ class CLargestRect {
             x0 = xw.first;
             w0 = xw.second;
 
-            if (width1 * (x - x0) > area(best_ll, best_ur)) {
+            if (width1*(x - x0) > area(best_ll, best_ur)) {
               best_ll = Point(x0, y - width1 + 1);
               best_ur = Point(x - 1, y);
             }
@@ -111,7 +113,7 @@ class CLargestRect {
  private:
   void updateCache(int y, std::vector<int> &c, const VALUE &value) const {
     for (int x = 0; x < width_; ++x) {
-      auto xi = size_t(x);
+      auto xi = uint(x);
 
       if (data_.getValue(x, y) == value)
         c[xi]++;
